crawler/requirements.txt
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
cffi==1.15.1
idna==3.4
kafka-python==2.0.2
pycparser==2.21
PySocks==1.7.1
selenium==3.141.0
sniffio
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
trio
trio-websocket
urllib3
wincertstore
wsproto
pytz
